Safari-611.1.10/LayoutTests/js/dom/unhandled-promise-rejection-handle-during-event-in-workers.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
worker = startWorker('../resources/unhandled-promise-rejection-handle-during-event-in-workers.js');
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
